Value + MultiAsset(/Mint) custom code
Based on old CML's Value / MultiAsset classes.

AssetBundle generalizes MultiAsset/Mint to work on any integer-like
value type.

Includes both checked and clamp operations, with the clamping/checking
applied uniformly to both the AssetBundle and Coin for Value, a
difference from the old Value where both used clamping operation for the
MultiAsset part.

Value was very lightly hand-edited to match closer the old format,
although without the multiasset part being an Option. This required one
extra field for ValueEncoding to differentiate between a coin-only Value
and one using the multiasset format but without any assets (e.g. empty
map).

Only includes the rust API. Once this is finalized we can work on the
wasm API.
