+++ chain/rust/src/serialization.rs
@@ -1946,238 +1946,3 @@ impl Deserialize for Update {
         .map_err(|e| e.annotate("Update"))
     }
 }
-
-impl Serialize for Value {
-    fn serialize<'se, W: Write>(
-        &self,
-        serializer: &'se mut Serializer<W>,
-        force_canonical: bool,
-    ) -> cbor_event::Result<&'se mut Serializer<W>> {
-        serializer.write_array_sz(
-            self.encodings
-                .as_ref()
-                .map(|encs| encs.len_encoding)
-                .unwrap_or_default()
-                .to_len_sz(2, force_canonical),
-        )?;
-        serializer.write_unsigned_integer_sz(
-            self.coin,
-            fit_sz(
-                self.coin,
-                self.encodings
-                    .as_ref()
-                    .map(|encs| encs.coin_encoding)
-                    .unwrap_or_default(),
-                force_canonical,
-            ),
-        )?;
-        serializer.write_map_sz(
-            self.encodings
-                .as_ref()
-                .map(|encs| encs.multiasset_encoding)
-                .unwrap_or_default()
-                .to_len_sz(self.multiasset.len() as u64, force_canonical),
-        )?;
-        let mut key_order = self
-            .multiasset
-            .iter()
-            .map(|(k, v)| {
-                let mut buf = cbor_event::se::Serializer::new_vec();
-                let multiasset_key_encoding = self
-                    .encodings
-                    .as_ref()
-                    .and_then(|encs| encs.multiasset_key_encodings.get(k))
-                    .cloned()
-                    .unwrap_or_default();
-                buf.write_bytes_sz(
-                    &k.to_raw_bytes(),
-                    multiasset_key_encoding
-                        .to_str_len_sz(k.to_raw_bytes().len() as u64, force_canonical),
-                )?;
-                Ok((buf.finalize(), k, v))
-            })
-            .collect::<Result<Vec<(Vec<u8>, &_, &_)>, cbor_event::Error>>()?;
-        if force_canonical {
-            key_order.sort_by(|(lhs_bytes, _, _), (rhs_bytes, _, _)| {
-                match lhs_bytes.len().cmp(&rhs_bytes.len()) {
-                    std::cmp::Ordering::Equal => lhs_bytes.cmp(rhs_bytes),
-                    diff_ord => diff_ord,
-                }
-            });
-        }
-        for (key_bytes, key, value) in key_order {
-            serializer.write_raw_bytes(&key_bytes)?;
-            let (multiasset_value_encoding, multiasset_value_value_encodings) = self
-                .encodings
-                .as_ref()
-                .and_then(|encs| encs.multiasset_value_encodings.get(key))
-                .cloned()
-                .unwrap_or_else(|| (LenEncoding::default(), BTreeMap::new()));
-            serializer.write_map_sz(
-                multiasset_value_encoding.to_len_sz(value.len() as u64, force_canonical),
-            )?;
-            let mut key_order = value
-                .iter()
-                .map(|(k, v)| {
-                    let mut buf = cbor_event::se::Serializer::new_vec();
-                    k.serialize(&mut buf, force_canonical)?;
-                    Ok((buf.finalize(), k, v))
-                })
-                .collect::<Result<Vec<(Vec<u8>, &_, &_)>, cbor_event::Error>>()?;
-            if force_canonical {
-                key_order.sort_by(|(lhs_bytes, _, _), (rhs_bytes, _, _)| {
-                    match lhs_bytes.len().cmp(&rhs_bytes.len()) {
-                        std::cmp::Ordering::Equal => lhs_bytes.cmp(rhs_bytes),
-                        diff_ord => diff_ord,
-                    }
-                });
-            }
-            for (key_bytes, key, value) in key_order {
-                serializer.write_raw_bytes(&key_bytes)?;
-                let multiasset_value_value_encoding = multiasset_value_value_encodings
-                    .get(key)
-                    .cloned()
-                    .unwrap_or_default();
-                serializer.write_unsigned_integer_sz(
-                    *value,
-                    fit_sz(*value, multiasset_value_value_encoding, force_canonical),
-                )?;
-            }
-            multiasset_value_encoding.end(serializer, force_canonical)?;
-        }
-        self.encodings
-            .as_ref()
-            .map(|encs| encs.multiasset_encoding)
-            .unwrap_or_default()
-            .end(serializer, force_canonical)?;
-        self.encodings
-            .as_ref()
-            .map(|encs| encs.len_encoding)
-            .unwrap_or_default()
-            .end(serializer, force_canonical)
-    }
-}
-
-impl Deserialize for Value {
-    fn deserialize<R: BufRead + Seek>(raw: &mut Deserializer<R>) -> Result<Self, DeserializeError> {
-        let len = raw.array_sz()?;
-        let len_encoding: LenEncoding = len.into();
-        let mut read_len = CBORReadLen::new(len);
-        read_len.read_elems(2)?;
-        (|| -> Result<_, DeserializeError> {
-            let (coin, coin_encoding) = raw
-                .unsigned_integer_sz()
-                .map(|(x, enc)| (x, Some(enc)))
-                .map_err(Into::<DeserializeError>::into)
-                .map_err(|e: DeserializeError| e.annotate("coin"))?;
-            let (
-                multiasset,
-                multiasset_encoding,
-                multiasset_key_encodings,
-                multiasset_value_encodings,
-            ) = (|| -> Result<_, DeserializeError> {
-                let mut multiasset_table = OrderedHashMap::new();
-                let multiasset_len = raw.map_sz()?;
-                let multiasset_encoding = multiasset_len.into();
-                let mut multiasset_key_encodings = BTreeMap::new();
-                let mut multiasset_value_encodings = BTreeMap::new();
-                while match multiasset_len {
-                    cbor_event::LenSz::Len(n, _) => (multiasset_table.len() as u64) < n,
-                    cbor_event::LenSz::Indefinite => true,
-                } {
-                    if raw.cbor_type()? == cbor_event::Type::Special {
-                        assert_eq!(raw.special()?, cbor_event::Special::Break);
-                        break;
-                    }
-                    let (multiasset_key, multiasset_key_encoding) = raw
-                        .bytes_sz()
-                        .map_err(Into::<DeserializeError>::into)
-                        .and_then(|(bytes, enc)| {
-                            ScriptHash::from_raw_bytes(&bytes)
-                                .map(|bytes| (bytes, StringEncoding::from(enc)))
-                                .map_err(|e| {
-                                    DeserializeFailure::InvalidStructure(Box::new(e)).into()
-                                })
-                        })?;
-                    let mut multiasset_value_table = OrderedHashMap::new();
-                    let multiasset_value_len = raw.map_sz()?;
-                    let multiasset_value_encoding = multiasset_value_len.into();
-                    let mut multiasset_value_value_encodings = BTreeMap::new();
-                    while match multiasset_value_len {
-                        cbor_event::LenSz::Len(n, _) => (multiasset_value_table.len() as u64) < n,
-                        cbor_event::LenSz::Indefinite => true,
-                    } {
-                        if raw.cbor_type()? == cbor_event::Type::Special {
-                            assert_eq!(raw.special()?, cbor_event::Special::Break);
-                            break;
-                        }
-                        let multiasset_value_key = AssetName::deserialize(raw)?;
-                        let (multiasset_value_value, multiasset_value_value_encoding) =
-                            raw.unsigned_integer_sz().map(|(x, enc)| (x, Some(enc)))?;
-                        if multiasset_value_table
-                            .insert(multiasset_value_key.clone(), multiasset_value_value)
-                            .is_some()
-                        {
-                            return Err(DeserializeFailure::DuplicateKey(Key::Str(String::from(
-                                "some complicated/unsupported type",
-                            )))
-                            .into());
-                        }
-                        multiasset_value_value_encodings
-                            .insert(multiasset_value_key, multiasset_value_value_encoding);
-                    }
-                    let (
-                        multiasset_value,
-                        multiasset_value_encoding,
-                        multiasset_value_value_encodings,
-                    ) = (
-                        multiasset_value_table,
-                        multiasset_value_encoding,
-                        multiasset_value_value_encodings,
-                    );
-                    if multiasset_table
-                        .insert(multiasset_key.clone(), multiasset_value)
-                        .is_some()
-                    {
-                        return Err(DeserializeFailure::DuplicateKey(Key::Str(String::from(
-                            "some complicated/unsupported type",
-                        )))
-                        .into());
-                    }
-                    multiasset_key_encodings
-                        .insert(multiasset_key.clone(), multiasset_key_encoding);
-                    multiasset_value_encodings.insert(
-                        multiasset_key.clone(),
-                        (multiasset_value_encoding, multiasset_value_value_encodings),
-                    );
-                }
-                Ok((
-                    multiasset_table,
-                    multiasset_encoding,
-                    multiasset_key_encodings,
-                    multiasset_value_encodings,
-                ))
-            })()
-            .map_err(|e| e.annotate("multiasset"))?;
-            match len {
-                cbor_event::LenSz::Len(_, _) => (),
-                cbor_event::LenSz::Indefinite => match raw.special()? {
-                    cbor_event::Special::Break => (),
-                    _ => return Err(DeserializeFailure::EndingBreakMissing.into()),
-                },
-            }
-            Ok(Value {
-                coin,
-                multiasset,
-                encodings: Some(ValueEncoding {
-                    len_encoding,
-                    coin_encoding,
-                    multiasset_encoding,
-                    multiasset_key_encodings,
-                    multiasset_value_encodings,
-                }),
-            })
-        })()
-        .map_err(|e| e.annotate("Value"))
-    }
-}
